[SONARCLOUD] Refactoring package "outils" and "vues"

+++ app/src/main/java/miage/fr/gestionprojet/outils/Outils.java
@@ -3,13 +3,9 @@ package miage.fr.gestionprojet.outils;
 import java.util.Calendar;
 import java.util.Date;
 
-/**
- * Created by Audrey on 27/02/2017.
- */
-
 public class Outils {
 
-    private static final long CONST_DURATION_OF_DAY = 1000l * 60 * 60 * 24;
+    private static final long CONST_DURATION_OF_DAY = (long) 1000 * 60 * 60 * 24;
 
     private Outils() {
         // private constructor
